Rename cli2 to cli
This moves all of the `cli2` command files into the `cli` directory,
since we are making the instance CLI the canonical CLI.

Test plan:
`yarn test --full` passes; `git grep cli2` returns no hits.
