+++ src/cli/bundledPlugins.js
