+++ src/cli/cliPlugin.js
