+++ src/cli/instanceConfig.js
