+++ src/cli/main.js
