+++ src/cli/score.js
