+++ src/cli/sourcecred.js
