--- src/cli2/command.js
@@ -1,3 +1,0 @@
-// @flow
-
-export * from "../cli/command";
