+++ src/plugins/discourse/cliPlugin.js
@@ -4,7 +4,7 @@ import Database from "better-sqlite3";
 import fs from "fs-extra";
 import {join as pathJoin} from "path";
 
-import type {CliPlugin, PluginDirectoryContext} from "../../cli2/cliPlugin";
+import type {CliPlugin, PluginDirectoryContext} from "../../cli/cliPlugin";
 import type {PluginDeclaration} from "../../analysis/pluginDeclaration";
 import type {ReferenceDetector} from "../../core/references/referenceDetector";
 import type {WeightedGraph} from "../../core/weightedGraph";
